pyproject.toml
[tool.poetry]
name = "handler"
version = "0.1.0"
description = ""
authors = ["II9tka <fffxx@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2"
djangorestframework = "^3.13.1"
pytracking = "^0.2.2"
Markdown = "^3.3.7"
django-ipware = "^4.0.2"
requests = "^2.27.1"
django-solo = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

